chore: move manifest reading to a separate package
