+++ internal/manifest/manifest_test.go
@@ -4,175 +4,15 @@ import (
 	"bytes"
 	"io"
 	"io/fs"
-	"os"
-	"path"
-	"slices"
-	"strings"
 
 	. "gopkg.in/check.v1"
 
 	"github.com/canonical/chisel/internal/archive"
-	"github.com/canonical/chisel/internal/manifest"
+	imanifest "github.com/canonical/chisel/internal/manifest"
 	"github.com/canonical/chisel/internal/setup"
+	"github.com/canonical/chisel/internal/testutil"
+	"github.com/canonical/chisel/manifest"
 )
-
-type manifestContents struct {
-	Paths    []*manifest.Path
-	Packages []*manifest.Package
-	Slices   []*manifest.Slice
-	Contents []*manifest.Content
-}
-
-var readManifestTests = []struct {
-	summary   string
-	input     string
-	mfest     *manifestContents
-	valError  string
-	readError string
-}{{
-	summary: "All types",
-	input: `
-		{"jsonwall":"1.0","schema":"1.0","count":13}
-		{"kind":"content","slice":"pkg1_manifest","path":"/manifest/manifest.wall"}
-		{"kind":"content","slice":"pkg1_myslice","path":"/dir/file"}
-		{"kind":"content","slice":"pkg1_myslice","path":"/dir/foo/bar/"}
-		{"kind":"content","slice":"pkg1_myslice","path":"/dir/link/file"}
-		{"kind":"content","slice":"pkg2_myotherslice","path":"/dir/foo/bar/"}
-		{"kind":"package","name":"pkg1","version":"v1","sha256":"hash1","arch":"arch1"}
-		{"kind":"package","name":"pkg2","version":"v2","sha256":"hash2","arch":"arch2"}
-		{"kind":"path","path":"/dir/file","mode":"0644","slices":["pkg1_myslice"],"sha256":"e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855","final_sha256":"8067926c032c090867013d14fb0eb21ae858344f62ad07086fd32375845c91a6","size":21}
-		{"kind":"path","path":"/dir/foo/bar/","mode":"01777","slices":["pkg2_myotherslice","pkg1_myslice"]}
-		{"kind":"path","path":"/dir/link/file","mode":"0644","slices":["pkg1_myslice"],"link":"/dir/file"}
-		{"kind":"path","path":"/manifest/manifest.wall","mode":"0644","slices":["pkg1_manifest"]}
-		{"kind":"slice","name":"pkg1_manifest"}
-		{"kind":"slice","name":"pkg1_myslice"}
-		{"kind":"slice","name":"pkg2_myotherslice"}
-	`,
-	mfest: &manifestContents{
-		Paths: []*manifest.Path{
-			{Kind: "path", Path: "/dir/file", Mode: "0644", Slices: []string{"pkg1_myslice"}, SHA256: "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855", FinalSHA256: "8067926c032c090867013d14fb0eb21ae858344f62ad07086fd32375845c91a6", Size: 0x15, Link: ""},
-			{Kind: "path", Path: "/dir/foo/bar/", Mode: "01777", Slices: []string{"pkg2_myotherslice", "pkg1_myslice"}, SHA256: "", FinalSHA256: "", Size: 0x0, Link: ""},
-			{Kind: "path", Path: "/dir/link/file", Mode: "0644", Slices: []string{"pkg1_myslice"}, SHA256: "", FinalSHA256: "", Size: 0x0, Link: "/dir/file"},
-			{Kind: "path", Path: "/manifest/manifest.wall", Mode: "0644", Slices: []string{"pkg1_manifest"}, SHA256: "", FinalSHA256: "", Size: 0x0, Link: ""},
-		},
-		Packages: []*manifest.Package{
-			{Kind: "package", Name: "pkg1", Version: "v1", Digest: "hash1", Arch: "arch1"},
-			{Kind: "package", Name: "pkg2", Version: "v2", Digest: "hash2", Arch: "arch2"},
-		},
-		Slices: []*manifest.Slice{
-			{Kind: "slice", Name: "pkg1_manifest"},
-			{Kind: "slice", Name: "pkg1_myslice"},
-			{Kind: "slice", Name: "pkg2_myotherslice"},
-		},
-		Contents: []*manifest.Content{
-			{Kind: "content", Slice: "pkg1_manifest", Path: "/manifest/manifest.wall"},
-			{Kind: "content", Slice: "pkg1_myslice", Path: "/dir/file"},
-			{Kind: "content", Slice: "pkg1_myslice", Path: "/dir/foo/bar/"},
-			{Kind: "content", Slice: "pkg1_myslice", Path: "/dir/link/file"},
-			{Kind: "content", Slice: "pkg2_myotherslice", Path: "/dir/foo/bar/"},
-		},
-	},
-}, {
-	summary: "Slice not found",
-	input: `
-		{"jsonwall":"1.0","schema":"1.0","count":1}
-		{"kind":"content","slice":"pkg1_manifest","path":"/manifest/manifest.wall"}
-	`,
-	valError: `invalid manifest: content path "/manifest/manifest.wall" refers to missing slice pkg1_manifest`,
-}, {
-	summary: "Package not found",
-	input: `
-		{"jsonwall":"1.0","schema":"1.0","count":1}
-		{"kind":"slice","name":"pkg1_manifest"}
-	`,
-	valError: `invalid manifest: slice pkg1_manifest refers to missing package "pkg1"`,
-}, {
-	summary: "Path not found in contents",
-	input: `
-		{"jsonwall":"1.0","schema":"1.0","count":1}
-		{"kind":"path","path":"/dir/","mode":"01777","slices":["pkg1_myslice"]}
-	`,
-	valError: `invalid manifest: path /dir/ has no matching entry in contents`,
-}, {
-	summary: "Content and path have different slices",
-	input: `
-		{"jsonwall":"1.0","schema":"1.0","count":3}
-		{"kind":"content","slice":"pkg1_myotherslice","path":"/dir/"}
-		{"kind":"package","name":"pkg1","version":"v1","sha256":"hash1","arch":"arch1"}
-		{"kind":"path","path":"/dir/","mode":"01777","slices":["pkg1_myslice"]}
-		{"kind":"slice","name":"pkg1_myotherslice"}
-	`,
-	valError: `invalid manifest: path /dir/ and content have diverging slices: \["pkg1_myslice"\] != \["pkg1_myotherslice"\]`,
-}, {
-	summary: "Content not found in paths",
-	input: `
-		{"jsonwall":"1.0","schema":"1.0","count":3}
-		{"kind":"content","slice":"pkg1_myslice","path":"/dir/"}
-		{"kind":"package","name":"pkg1","version":"v1","sha256":"hash1","arch":"arch1"}
-		{"kind":"slice","name":"pkg1_myslice"}
-	`,
-	valError: `invalid manifest: content path /dir/ has no matching entry in paths`,
-}, {
-	summary: "Malformed jsonwall",
-	input: `
-		{"jsonwall":"1.0","schema":"1.0","count":1}
-		{"kind":"content", "not valid json"
-	`,
-	valError: `invalid manifest: cannot read manifest: unexpected end of JSON input`,
-}, {
-	summary: "Unknown schema",
-	input: `
-		{"jsonwall":"1.0","schema":"2.0","count":1}
-		{"kind":"package","name":"pkg1","version":"v1","sha256":"hash1","arch":"arch1"}
-	`,
-	readError: `cannot read manifest: unknown schema version "2.0"`,
-}}
-
-func (s *S) TestManifestReadValidate(c *C) {
-	for _, test := range readManifestTests {
-		c.Logf("Summary: %s", test.summary)
-
-		// Reindent the jsonwall to remove leading tabs in each line.
-		lines := strings.Split(strings.TrimSpace(test.input), "\n")
-		trimmedLines := make([]string, 0, len(lines))
-		for _, line := range lines {
-			trimmedLines = append(trimmedLines, strings.TrimLeft(line, "\t"))
-		}
-		test.input = strings.Join(trimmedLines, "\n")
-		// Assert that the jsonwall is valid, for the test to be meaningful.
-		slices.Sort(trimmedLines)
-		orderedInput := strings.Join(trimmedLines, "\n")
-		c.Assert(test.input, DeepEquals, orderedInput, Commentf("input jsonwall lines should be ordered"))
-
-		tmpDir := c.MkDir()
-		manifestPath := path.Join(tmpDir, "manifest.wall")
-		w, err := os.OpenFile(manifestPath, os.O_CREATE|os.O_TRUNC|os.O_WRONLY, 0644)
-		c.Assert(err, IsNil)
-		_, err = w.Write([]byte(test.input))
-		c.Assert(err, IsNil)
-		w.Close()
-
-		r, err := os.OpenFile(manifestPath, os.O_RDONLY, 0644)
-		c.Assert(err, IsNil)
-		defer r.Close()
-
-		mfest, err := manifest.Read(r)
-		if test.readError != "" {
-			c.Assert(err, ErrorMatches, test.readError)
-			continue
-		}
-		c.Assert(err, IsNil)
-		err = manifest.Validate(mfest)
-		if test.valError != "" {
-			c.Assert(err, ErrorMatches, test.valError)
-			continue
-		}
-		c.Assert(err, IsNil)
-		if test.mfest != nil {
-			c.Assert(dumpManifestContents(c, mfest), DeepEquals, test.mfest)
-		}
-	}
-}
 
 var findPathsTests = []struct {
 	summary  string
@@ -247,7 +87,7 @@ func (s *S) TestFindPaths(c *C) {
 	for _, test := range findPathsTests {
 		c.Logf("Summary: %s", test.summary)
 
-		manifestSlices := manifest.FindPaths(test.slices)
+		manifestSlices := imanifest.FindPaths(test.slices)
 
 		slicesByName := map[string]*setup.Slice{}
 		for _, slice := range test.slices {
@@ -277,17 +117,17 @@ var slice2 = &setup.Slice{
 
 var generateManifestTests = []struct {
 	summary     string
-	report      *manifest.Report
+	report      *imanifest.Report
 	packageInfo []*archive.PackageInfo
 	selection   []*setup.Slice
-	expected    *manifestContents
+	expected    *testutil.ManifestContents
 	error       string
 }{{
 	summary:   "Basic",
 	selection: []*setup.Slice{slice1, slice2},
-	report: &manifest.Report{
+	report: &imanifest.Report{
 		Root: "/",
-		Entries: map[string]manifest.ReportEntry{
+		Entries: map[string]imanifest.ReportEntry{
 			"/file": {
 				Path:        "/file",
 				Mode:        0456,
@@ -315,7 +155,7 @@ var generateManifestTests = []struct {
 		Arch:    "a2",
 		SHA256:  "s2",
 	}},
-	expected: &manifestContents{
+	expected: &testutil.ManifestContents{
 		Paths: []*manifest.Path{{
 			Kind:        "path",
 			Path:        "/file",
@@ -367,9 +207,9 @@ var generateManifestTests = []struct {
 	},
 }, {
 	summary: "Missing slice",
-	report: &manifest.Report{
+	report: &imanifest.Report{
 		Root: "/",
-		Entries: map[string]manifest.ReportEntry{
+		Entries: map[string]imanifest.ReportEntry{
 			"/file": {
 				Path:        "/file",
 				Mode:        0456,
@@ -384,9 +224,9 @@ var generateManifestTests = []struct {
 	error:     `internal error: invalid manifest: path "/file" refers to missing slice package1_slice1`,
 }, {
 	summary: "Missing package",
-	report: &manifest.Report{
+	report: &imanifest.Report{
 		Root: "/",
-		Entries: map[string]manifest.ReportEntry{
+		Entries: map[string]imanifest.ReportEntry{
 			"/file": {
 				Path:        "/file",
 				Mode:        0456,
@@ -401,9 +241,9 @@ var generateManifestTests = []struct {
 	error:       `internal error: invalid manifest: slice package1_slice1 refers to missing package "package1"`,
 }, {
 	summary: "Invalid path: link set for regular file",
-	report: &manifest.Report{
+	report: &imanifest.Report{
 		Root: "/",
-		Entries: map[string]manifest.ReportEntry{
+		Entries: map[string]imanifest.ReportEntry{
 			"/file": {
 				Path:   "/file",
 				Mode:   0456,
@@ -415,9 +255,9 @@ var generateManifestTests = []struct {
 	error: `internal error: invalid manifest: path "/file" has invalid options: link set for regular file`,
 }, {
 	summary: "Invalid path: slices is empty",
-	report: &manifest.Report{
+	report: &imanifest.Report{
 		Root: "/",
-		Entries: map[string]manifest.ReportEntry{
+		Entries: map[string]imanifest.ReportEntry{
 			"/file": {
 				Path: "/file",
 				Mode: 0456,
@@ -427,9 +267,9 @@ var generateManifestTests = []struct {
 	error: `internal error: invalid manifest: path "/file" has invalid options: slices is empty`,
 }, {
 	summary: "Invalid path: link set for symlink",
-	report: &manifest.Report{
+	report: &imanifest.Report{
 		Root: "/",
-		Entries: map[string]manifest.ReportEntry{
+		Entries: map[string]imanifest.ReportEntry{
 			"/link": {
 				Path:   "/link",
 				Mode:   0456 | fs.ModeSymlink,
@@ -440,9 +280,9 @@ var generateManifestTests = []struct {
 	error: `internal error: invalid manifest: path "/link" has invalid options: link not set for symlink`,
 }, {
 	summary: "Invalid path: sha256 set for symlink",
-	report: &manifest.Report{
+	report: &imanifest.Report{
 		Root: "/",
-		Entries: map[string]manifest.ReportEntry{
+		Entries: map[string]imanifest.ReportEntry{
 			"/link": {
 				Path:   "/link",
 				Mode:   0456 | fs.ModeSymlink,
@@ -455,9 +295,9 @@ var generateManifestTests = []struct {
 	error: `internal error: invalid manifest: path "/link" has invalid options: sha256 set for symlink`,
 }, {
 	summary: "Invalid path: final_sha256 set for symlink",
-	report: &manifest.Report{
+	report: &imanifest.Report{
 		Root: "/",
-		Entries: map[string]manifest.ReportEntry{
+		Entries: map[string]imanifest.ReportEntry{
 			"/link": {
 				Path:        "/link",
 				Mode:        0456 | fs.ModeSymlink,
@@ -470,9 +310,9 @@ var generateManifestTests = []struct {
 	error: `internal error: invalid manifest: path "/link" has invalid options: final_sha256 set for symlink`,
 }, {
 	summary: "Invalid path: size set for symlink",
-	report: &manifest.Report{
+	report: &imanifest.Report{
 		Root: "/",
-		Entries: map[string]manifest.ReportEntry{
+		Entries: map[string]imanifest.ReportEntry{
 			"/link": {
 				Path:   "/link",
 				Mode:   0456 | fs.ModeSymlink,
@@ -485,9 +325,9 @@ var generateManifestTests = []struct {
 	error: `internal error: invalid manifest: path "/link" has invalid options: size set for symlink`,
 }, {
 	summary: "Invalid path: link set for directory",
-	report: &manifest.Report{
+	report: &imanifest.Report{
 		Root: "/",
-		Entries: map[string]manifest.ReportEntry{
+		Entries: map[string]imanifest.ReportEntry{
 			"/dir": {
 				Path:   "/dir",
 				Mode:   0456 | fs.ModeDir,
@@ -499,9 +339,9 @@ var generateManifestTests = []struct {
 	error: `internal error: invalid manifest: path "/dir" has invalid options: link set for directory`,
 }, {
 	summary: "Invalid path: sha256 set for directory",
-	report: &manifest.Report{
+	report: &imanifest.Report{
 		Root: "/",
-		Entries: map[string]manifest.ReportEntry{
+		Entries: map[string]imanifest.ReportEntry{
 			"/dir": {
 				Path:   "/dir",
 				Mode:   0456 | fs.ModeDir,
@@ -513,9 +353,9 @@ var generateManifestTests = []struct {
 	error: `internal error: invalid manifest: path "/dir" has invalid options: sha256 set for directory`,
 }, {
 	summary: "Invalid path: final_sha256 set for directory",
-	report: &manifest.Report{
+	report: &imanifest.Report{
 		Root: "/",
-		Entries: map[string]manifest.ReportEntry{
+		Entries: map[string]imanifest.ReportEntry{
 			"/dir": {
 				Path:        "/dir",
 				Mode:        0456 | fs.ModeDir,
@@ -527,9 +367,9 @@ var generateManifestTests = []struct {
 	error: `internal error: invalid manifest: path "/dir" has invalid options: final_sha256 set for directory`,
 }, {
 	summary: "Invalid path: size set for directory",
-	report: &manifest.Report{
+	report: &imanifest.Report{
 		Root: "/",
-		Entries: map[string]manifest.ReportEntry{
+		Entries: map[string]imanifest.ReportEntry{
 			"/dir": {
 				Path:   "/dir",
 				Mode:   0456 | fs.ModeDir,
@@ -588,13 +428,13 @@ func (s *S) TestGenerateManifests(c *C) {
 			}}
 		}
 
-		options := &manifest.WriteOptions{
+		options := &imanifest.WriteOptions{
 			PackageInfo: test.packageInfo,
 			Selection:   test.selection,
 			Report:      test.report,
 		}
 		var buffer bytes.Buffer
-		err := manifest.Write(options, &buffer)
+		err := imanifest.Write(options, &buffer)
 		if test.error != "" {
 			c.Assert(err, ErrorMatches, test.error)
 			continue
@@ -604,19 +444,19 @@ func (s *S) TestGenerateManifests(c *C) {
 		c.Assert(err, IsNil)
 		err = manifest.Validate(mfest)
 		c.Assert(err, IsNil)
-		contents := dumpManifestContents(c, mfest)
+		contents := testutil.DumpManifestContents(c, mfest)
 		c.Assert(contents, DeepEquals, test.expected)
 	}
 }
 
 func (s *S) TestGenerateNoManifests(c *C) {
-	report, err := manifest.NewReport("/")
+	report, err := imanifest.NewReport("/")
 	c.Assert(err, IsNil)
-	options := &manifest.WriteOptions{
+	options := &imanifest.WriteOptions{
 		Report: report,
 	}
 	var buffer bytes.Buffer
-	err = manifest.Write(options, &buffer)
+	err = imanifest.Write(options, &buffer)
 	c.Assert(err, IsNil)
 
 	var reader io.Reader = &buffer
@@ -624,42 +464,4 @@ func (s *S) TestGenerateNoManifests(c *C) {
 	n, err := reader.Read(bs)
 	c.Assert(err, IsNil)
 	c.Assert(n, Equals, 0)
-}
-
-func dumpManifestContents(c *C, mfest *manifest.Manifest) *manifestContents {
-	var slices []*manifest.Slice
-	err := mfest.IterateSlices("", func(slice *manifest.Slice) error {
-		slices = append(slices, slice)
-		return nil
-	})
-	c.Assert(err, IsNil)
-
-	var pkgs []*manifest.Package
-	err = mfest.IteratePackages(func(pkg *manifest.Package) error {
-		pkgs = append(pkgs, pkg)
-		return nil
-	})
-	c.Assert(err, IsNil)
-
-	var paths []*manifest.Path
-	err = mfest.IteratePaths("", func(path *manifest.Path) error {
-		paths = append(paths, path)
-		return nil
-	})
-	c.Assert(err, IsNil)
-
-	var contents []*manifest.Content
-	err = mfest.IterateContents("", func(content *manifest.Content) error {
-		contents = append(contents, content)
-		return nil
-	})
-	c.Assert(err, IsNil)
-
-	mc := manifestContents{
-		Paths:    paths,
-		Packages: pkgs,
-		Slices:   slices,
-		Contents: contents,
-	}
-	return &mc
 }
